requirements.txt
librosa==0.9.2
numpy==2.0.1
opencv_python==4.7.0.72
Pillow==10.4.0
sounddevice==0.4.7
streamlit==1.36.0
streamlit_webrtc==0.47.7
tensorflow==2.17.0
tensorflow_intel==2.17.0
